Fix AwMediaUrlInterceptorTest.TestInterceptInvalidAssetUrl test.
Fix the AwMediaUrlInterceptorTest.TestInterceptInvalidAssetUrl
test which is failing on a number of bots.

This is due to https://ccrev.com/c/761408 where size was changed
to be of type size_t.

BUG=784572

+++ android_webview/browser/aw_media_url_interceptor.cc
@@ -28,7 +28,7 @@ bool AwMediaUrlInterceptor::Intercept(const std::string& url,
         base::MemoryMappedFile::Region::kWholeFile;
     *fd = base::android::OpenApkAsset(filename, &region);
     *offset = region.offset;
-    *size = region.size;
+    *size = static_cast<int64_t>(region.size);
     return *fd != -1;
   }
